+++ docs/src/components/ColorVariants.tsx
@@ -6,6 +6,8 @@ import ColorBox from './ColorBox'
 
 type ColorVariantsProps = {
   name: string
+  headerOnly?: boolean
+  size?: 'small' | 'normal' | 'big'
 }
 
 export const colorVariantsStyles: ComponentSlotStylesInput<ColorVariantsProps> = {
@@ -18,19 +20,30 @@ export const colorVariantsStyles: ComponentSlotStylesInput<ColorVariantsProps> =
 
 const ColorVariants = createComponent<ColorVariantsProps>({
   displayName: 'ColorVariants',
-  render: ({ name, stardust: { classes } }) => (
+  render: ({ name, headerOnly, size, stardust: { classes } }) => (
     <ProviderConsumer
       render={({ siteVariables: { colors } }) => (
         <div className={classes.root}>
-          <ColorBox name={name} size="big" value={colors[name][500]} copyToClipboardIcon={false} />
+          <ColorBox
+            name={name}
+            size={size}
+            value={colors[name][600] || colors[name][500]}
+            copyToClipboardIcon={false}
+          />
 
-          {_.map(colors[name], (value, variable) => (
-            <ColorBox key={variable} name={variable} size="small" value={value} />
-          ))}
+          {!headerOnly &&
+            _.map(colors[name], (value, variable) => (
+              <ColorBox key={variable} name={variable} size="small" value={value} />
+            ))}
         </div>
       )}
     />
   ),
 })
+
+ColorVariants.defaultProps = {
+  headerOnly: false,
+  size: 'big',
+}
 
 export default ColorVariants